@@ -5,7 +5,7 @@ publish_to: none
 homepage: https://github.com/jamiewest/signalr_core
 
 environment:
-  sdk: '>=2.7.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   http: ^0.13.0
