@@ -9,11 +9,11 @@ import 'package:signalr_core/src/retry_policy.dart';
 import 'package:tuple/tuple.dart';
 
 typedef InvocationEventCallback = void Function(
-    HubMessage invocationEvent, Exception exception);
-typedef MethodInvacationFunc = void Function(List<dynamic> arguments);
-typedef ClosedCallback = void Function(Exception exception);
-typedef ReconnectingCallback = void Function(Exception exception);
-typedef ReconnectedCallback = void Function(String connectionId);
+    HubMessage? invocationEvent, Exception? exception);
+typedef MethodInvocationFunc = void Function(List<dynamic>? arguments);
+typedef ClosedCallback = void Function(Exception? exception);
+typedef ReconnectingCallback = void Function(Exception? exception);
+typedef ReconnectedCallback = void Function(String? connectionId);
 
 const int DEFAULT_TIMEOUT_IN_MS = 30 * 1000;
 const int DEFAULT_PING_INTERVAL_IN_MS = 15 * 1000;
@@ -39,37 +39,37 @@ enum HubConnectionState {
 /// Represents a connection to a SignalR Hub.
 class HubConnection {
   final dynamic _cachedPingMessage;
-  final Connection _connection;
-  final Logging _logger;
-  final RetryPolicy _reconnectPolicy;
+  final Connection? _connection;
+  final Logging? _logger;
+  final RetryPolicy? _reconnectPolicy;
   final HubProtocol _protocol;
-  HandshakeProtocol _handshakeProtocol;
-  Map<String, InvocationEventCallback> _callbacks;
-  Map<String, List<MethodInvacationFunc>> _methods;
-  int _invocationId;
+  late HandshakeProtocol _handshakeProtocol;
+  late Map<String?, InvocationEventCallback> _callbacks;
+  late Map<String, List<MethodInvocationFunc>> _methods;
+  int? _invocationId;
 
-  List<ClosedCallback> _closedCallbacks;
-  List<ReconnectingCallback> _reconnectingCallbacks;
-  List<ReconnectedCallback> _reconnectedCallbacks;
+  late List<ClosedCallback> _closedCallbacks;
+  late List<ReconnectingCallback> _reconnectingCallbacks;
+  late List<ReconnectedCallback> _reconnectedCallbacks;
 
-  bool _receivedHandshakeResponse;
-  Completer _handshakeCompleter;
-  Exception _stopDuringStartError;
+  bool? _receivedHandshakeResponse;
+  late Completer _handshakeCompleter;
+  Exception? _stopDuringStartError;
 
-  HubConnectionState _connectionState;
-  bool _connectionStarted;
-  Future<void> _startFuture;
-  Future<void> _stopFuture;
+  HubConnectionState? _connectionState;
+  late bool _connectionStarted;
+  Future<void>? _startFuture;
+  Future<void>? _stopFuture;
 
-  Timer _reconnectDelayHandle;
-  Timer _timeoutHandle;
-  Timer _pingServerHandle;
+  Timer? _reconnectDelayHandle;
+  Timer? _timeoutHandle;
+  Timer? _pingServerHandle;
 
   HubConnection({
-    Connection connection,
-    Logging logging,
-    HubProtocol protocol,
-    RetryPolicy reconnectPolicy,
+    Connection? connection,
+    Logging? logging,
+    required HubProtocol protocol,
+    RetryPolicy? reconnectPolicy,
   })  : _cachedPingMessage = protocol.writeMessage(PingMessage()),
         _connection = connection,
         _logger = logging,
@@ -79,10 +79,12 @@ class HubConnection {
     keepAliveIntervalInMilliseconds = DEFAULT_PING_INTERVAL_IN_MS;
 
     _handshakeProtocol = HandshakeProtocol();
-    _connection
-      ..onreceive = _processIncomingData
-      ..onclose =
-          (Exception exception) => _connectionClosed(exception: exception);
+    if (_connection != null) {
+      _connection!
+        ..onreceive = _processIncomingData
+        ..onclose =
+            (Exception? exception) => _connectionClosed(exception: exception);
+    }
 
     _callbacks = {};
     _methods = {};
@@ -99,23 +101,23 @@ class HubConnection {
   ///
   /// If this timeout elapses without receiving any messages from the server, the connection will be terminated with an error.
   /// The default timeout value is 30,000 milliseconds (30 seconds).
-  int serverTimeoutInMilliseconds;
+  late int serverTimeoutInMilliseconds;
 
   /// Default interval at which to ping the server.
   ///
   /// The default value is 15,000 milliseconds (15 seconds).
   /// Allows the server to detect hard disconnects (like when a client unplugs their computer).
-  int keepAliveIntervalInMilliseconds;
+  late int keepAliveIntervalInMilliseconds;
 
   /// Indicates the state of the {@link HubConnection} to the server.
-  HubConnectionState get state => _connectionState;
+  HubConnectionState? get state => _connectionState;
 
   /// Represents the connection id of the [HubConnection] on the server. The connection id will be null when the connection is either
   /// in the disconnected state or if the negotiation step was skipped.
-  String get connectionId => _connection.connectionId;
+  String? get connectionId => _connection!.connectionId;
 
   /// Indicates the url of the {@link HubConnection} to the server.
-  String get baseUrl => _connection.baseUrl;
+  String get baseUrl => _connection!.baseUrl!;
 
   /// Sets a new url for the HubConnection. Note that the url can only be changed when the connection is in either the Disconnected or
   /// Reconnecting states.
@@ -130,11 +132,11 @@ class HubConnection {
       throw Exception('The HubConnection url must be a valid url.');
     }
 
-    _connection.baseUrl = url;
+    _connection!.baseUrl = url;
   }
 
   /// Starts the connection.
-  Future<void> start() {
+  Future<void>? start() {
     _startFuture = _startWithStateTransitions();
     return _startFuture;
   }
@@ -146,17 +148,17 @@ class HubConnection {
     }
 
     _connectionState = HubConnectionState.connecting;
-    _logger(LogLevel.debug, 'Starting HubConnection.');
+    _logger!(LogLevel.debug, 'Starting HubConnection.');
 
     try {
       await _startInternal();
 
       _connectionState = HubConnectionState.connected;
       _connectionStarted = true;
-      _logger(LogLevel.debug, 'HubConnection connected successfully.');
+      _logger!(LogLevel.debug, 'HubConnection connected successfully.');
     } catch (e) {
       _connectionState = HubConnectionState.disconnected;
-      _logger(LogLevel.debug,
+      _logger!(LogLevel.debug,
           'HubConnection failed to start successfully because of error \'{$e.toString()}\'.');
       return Future.error(e);
     }
@@ -175,7 +177,7 @@ class HubConnection {
     //   _handshakeRejector = reject
     // });
 
-    await _connection.start(transferFormat: _protocol.transferFormat);
+    await _connection!.start(transferFormat: _protocol.transferFormat);
 
     try {
       final handshakeRequest = HandshakeRequestMessage(
@@ -183,12 +185,13 @@ class HubConnection {
         version: _protocol.version,
       );
 
-      _logger(LogLevel.debug, 'Sending handshake request.');
+      _logger!(LogLevel.debug, 'Sending handshake request.');
 
       await _sendMessage(
           _handshakeProtocol.writeHandshakeRequest(handshakeRequest));
 
-      _logger(LogLevel.information, 'Using HubProtocol \'${_protocol.name}\'.');
+      _logger!(
+          LogLevel.information, 'Using HubProtocol \'${_protocol.name}\'.');
 
       // defensively cleanup timeout in case we receive a message from the server before we finish start
       _cleanupTimeout();
@@ -204,10 +207,10 @@ class HubConnection {
         // It's important to throw instead of returning a rejected promise, because we don't want to allow any state
         // transitions to occur between now and the calling code observing the exceptions. Returning a rejected promise
         // will cause the calling continuation to get scheduled to run later.
-        throw _stopDuringStartError;
+        throw _stopDuringStartError!;
       }
     } catch (e) {
-      _logger(LogLevel.debug,
+      _logger!(LogLevel.debug,
           'Hub handshake failed with error \'${e.toString()}\' during start(). Stopping HubConnection.');
 
       _cleanupTimeout();
@@ -215,7 +218,7 @@ class HubConnection {
 
       // HttpConnection.stop() should not complete until after the onclose callback is invoked.
       // This will transition the HubConnection to the disconnected state before HttpConnection.stop() completes.
-      await _connection.stop(exception: e as Exception);
+      await _connection!.stop(exception: e as Exception);
       rethrow;
     }
   }
@@ -236,28 +239,28 @@ class HubConnection {
     }
   }
 
-  Future<void> _stopInternal({Exception exception}) async {
+  Future<void>? _stopInternal({Exception? exception}) async {
     if (_connectionState == HubConnectionState.disconnected) {
-      _logger(LogLevel.debug,
+      _logger!(LogLevel.debug,
           'Call to HubConnection.stop(${exception.toString()}) ignored because it is already in the disconnected state.');
       return Future.value(null);
     }
 
     if (_connectionState == HubConnectionState.disconnecting) {
-      _logger(LogLevel.debug,
+      _logger!(LogLevel.debug,
           'Call to HttpConnection.stop(${exception.toString()}) ignored because the connection is already in the disconnecting state.');
       return _stopFuture;
     }
 
     _connectionState = HubConnectionState.disconnecting;
 
-    _logger(LogLevel.debug, 'Stopping HubConnection');
+    _logger!(LogLevel.debug, 'Stopping HubConnection');
 
     if (_reconnectDelayHandle != null) {
       // We're in a reconnect delay which means the underlying connection is currently already stopped.
       // Just clear the handle to stop the reconnect loop (which no one is waiting on thankfully) and
       // fire the onclose callbacks.
-      _logger(LogLevel.debug,
+      _logger!(LogLevel.debug,
           'Connection stopped during reconnect delay. Done reconnecting.');
 
       _clearTimeout(_reconnectDelayHandle);
@@ -277,7 +280,7 @@ class HubConnection {
     // HttpConnection.stop() should not complete until after either HttpConnection.start() fails
     // or the onclose callback is invoked. The onclose callback will transition the HubConnection
     // to the disconnected state if need be before HttpConnection.stop() completes.
-    return _connection.stop(exception: exception);
+    return _connection!.stop(exception: exception);
   }
 
   /// Invokes a hub method on the server using the specified name and arguments. Does not wait for a response from the receiver.
@@ -285,8 +288,8 @@ class HubConnection {
   /// The Promise returned by this method resolves when the client has sent the invocation to the server. The server may still
   /// be processing the invocation.
   Future<void> send({
-    String methodName,
-    List<dynamic> args,
+    String? methodName,
+    List<dynamic>? args,
   }) {
     final streamParameters = _replaceStreamParameters(args);
     final sendPromise = _sendWithProtocol(
@@ -305,7 +308,7 @@ class HubConnection {
 
   Future<void> _sendMessage(dynamic message) {
     _resetKeepAliveInterval();
-    return _connection.send(message);
+    return _connection!.send(message);
   }
 
   Future<void> _sendWithProtocol(dynamic message) {
@@ -334,8 +337,8 @@ class HubConnection {
   }
 
   void _resetTimeoutPeriod() {
-    if ((_connection.features != null) ||
-        (_connection.features.inherentKeepAlive == false)) {
+    if ((_connection!.features != null) ||
+        (_connection!.features.inherentKeepAlive == false)) {
       // Set the timeout timer
       _timeoutHandle = Timer.periodic(
           Duration(milliseconds: serverTimeoutInMilliseconds), (Timer timer) {
@@ -349,13 +352,13 @@ class HubConnection {
     // The server hasn't talked to us in a while. It doesn't like us anymore ... :(
     // Terminate the connection, but we don't need to wait on the promise. This could trigger reconnecting.
     // tslint:disable-next-line:no-floating-promises
-    _connection.stop(
+    _connection!.stop(
       exception: Exception(
           'Server timeout elapsed without receiving a message from the server.'),
     );
   }
 
-  void _completeClose({Exception exception}) {
+  void _completeClose({Exception? exception}) {
     if (_connectionStarted) {
       _connectionState = HubConnectionState.disconnected;
       _connectionStarted = false;
@@ -365,13 +368,13 @@ class HubConnection {
           callback(exception);
         }
       } catch (e) {
-        _logger(LogLevel.error,
+        _logger!(LogLevel.error,
             'An onclose callback called with error \'${exception.toString()}\' threw error \'${e.toString()}\'.');
       }
     }
   }
 
-  Future<void> _reconnect({Exception exception}) async {
+  Future<void> _reconnect({Exception? exception}) async {
     final reconnectStartTime = Stopwatch()..start();
     //final reconnectStartTime = DateTime.now();
     var previousReconnectAttempts = 0;
@@ -385,7 +388,7 @@ class HubConnection {
         retryReason: retryError);
 
     if (nextRetryDelay == null) {
-      _logger(LogLevel.debug,
+      _logger!(LogLevel.debug,
           'Connection not reconnecting because the RetryPolicy returned null on the first reconnect attempt.');
       _completeClose(exception: exception);
       return;
@@ -394,44 +397,42 @@ class HubConnection {
     _connectionState = HubConnectionState.reconnecting;
 
     if (exception != null) {
-      _logger(LogLevel.information,
+      _logger!(LogLevel.information,
           'Connection reconnecting because of error \'${exception.toString()}\'.');
     } else {
-      _logger(LogLevel.information, 'Connection reconnecting.');
+      _logger!(LogLevel.information, 'Connection reconnecting.');
     }
 
-    if (onreconnecting != null) {
-      try {
-        for (var callback in _reconnectingCallbacks) {
-          callback(exception);
-        }
-      } catch (e) {
-        _logger(LogLevel.error,
-            'An onreconnecting callback called with error \'${exception.toString()}\' threw error \'${e.toString()}\'.');
+    try {
+      for (var callback in _reconnectingCallbacks) {
+        callback(exception);
       }
+    } catch (e) {
+      _logger!(LogLevel.error,
+          'An onreconnecting callback called with error \'${exception.toString()}\' threw error \'${e.toString()}\'.');
+    }
 
-      // Exit early if an onreconnecting callback called connection.stop().
-      if (_connectionState != HubConnectionState.reconnecting) {
-        _logger(LogLevel.debug,
-            'Connection left the reconnecting state in onreconnecting callback. Done reconnecting.');
-        return;
-      }
+    // Exit early if an onreconnecting callback called connection.stop().
+    if (_connectionState != HubConnectionState.reconnecting) {
+      _logger!(LogLevel.debug,
+          'Connection left the reconnecting state in onreconnecting callback. Done reconnecting.');
+      return;
     }
 
     while (nextRetryDelay != null) {
-      _logger(LogLevel.information,
+      _logger!(LogLevel.information,
           'Reconnect attempt number $previousReconnectAttempts will start in $nextRetryDelay ms.');
 
       await Future(() {
         var completer = Completer();
-        _reconnectDelayHandle = Timer(
-            Duration(milliseconds: nextRetryDelay), () => completer.complete());
+        _reconnectDelayHandle = Timer(Duration(milliseconds: nextRetryDelay!),
+            () => completer.complete());
         return completer.future;
       });
       _reconnectDelayHandle = null;
 
-      if (_connectionState != HubConnectionState.reconnecting) {
-        _logger(LogLevel.debug,
+      if (_connectionState == null || _connectionState != HubConnectionState.reconnecting) {
+        _logger!(LogLevel.debug,
             'Connection left the reconnecting state during reconnect delay. Done reconnecting.');
         return;
       }
@@ -440,27 +441,25 @@ class HubConnection {
         await _startInternal();
 
         _connectionState = HubConnectionState.connected;
-        _logger(
+        _logger!(
             LogLevel.information, 'HubConnection reconnected successfully.');
 
-        if (onreconnected != null) {
-          try {
-            for (var callback in _reconnectedCallbacks) {
-              callback(_connection.connectionId);
-            }
-          } catch (e) {
-            _logger(LogLevel.error,
-                'An onreconnected callback called with connectionId \'${_connection.connectionId}; threw error \'${e.toString()}\'.');
+        try {
+          for (var callback in _reconnectedCallbacks) {
+            callback(_connection!.connectionId);
           }
+        } catch (e) {
+          _logger!(LogLevel.error,
+              'An onreconnected callback called with connectionId \'${_connection!.connectionId}; threw error \'${e.toString()}\'.');
         }
 
         return;
       } catch (e) {
-        _logger(LogLevel.information,
+        _logger!(LogLevel.information,
             'Reconnect attempt failed because of error \'${e.toString()}\'.');
 
         if (_connectionState != HubConnectionState.reconnecting) {
-          _logger(LogLevel.debug,
+          _logger!(LogLevel.debug,
               'Connection left the reconnecting state during reconnect attempt. Done reconnecting.');
           return;
         }
@@ -474,19 +473,19 @@ class HubConnection {
       }
     }
 
-    _logger(LogLevel.information,
+    _logger!(LogLevel.information,
         'Reconnect retries have been exhausted after ${reconnectStartTime.elapsedMilliseconds} ms and $previousReconnectAttempts failed attempts. Connection disconnecting.');
 
     _completeClose();
   }
 
-  int _getNextRetryDelay({
-    int previousRetryCount,
-    int elapsedMilliseconds,
-    Exception retryReason,
+  int? _getNextRetryDelay({
+    int? previousRetryCount,
+    int? elapsedMilliseconds,
+    Exception? retryReason,
   }) {
     try {
-      return _reconnectPolicy.nextRetryDelayInMilliseconds(
+      return _reconnectPolicy!.nextRetryDelayInMilliseconds(
         RetryContext(
           elapsedMilliseconds: elapsedMilliseconds,
           previousRetryCount: previousRetryCount,
@@ -494,20 +493,20 @@ class HubConnection {
         ),
       );
     } catch (e) {
-      _logger(LogLevel.error,
+      _logger!(LogLevel.error,
           'RetryPolicy.nextRetryDelayInMilliseconds($previousRetryCount, $elapsedMilliseconds) threw error \'${e.toString()}\'.');
       return null;
     }
   }
 
   /// Invokes a streaming hub method on the server using the specified name and arguments.
-  Stream<T> stream<T>(String methodName, {List<dynamic> args}) {
+  Stream<T?> stream<T>(String methodName, {List<dynamic>? args}) {
     final streamParameters = _replaceStreamParameters(args);
     final invocationDescriptor = _createStreamInvocation(
         methodName: methodName, args: args, streamIds: streamParameters.item2);
 
-    Future<void> futureQueue;
-    final controller = StreamController<T>()
+    late Future<void> futureQueue;
+    final controller = StreamController<T?>()
       ..onCancel = () {
         final cancelInvocation =
             _createCancelInvocation(id: invocationDescriptor.invocationId);
@@ -520,7 +519,7 @@ class HubConnection {
       };
 
     _callbacks[invocationDescriptor.invocationId] =
-        (HubMessage invocationEvent, Exception error) {
+        (HubMessage? invocationEvent, Exception? error) {
       if (error != null) {
         controller.addError(error);
         return;
@@ -528,7 +527,7 @@ class HubConnection {
         if (invocationEvent.type == MessageType.completion) {
           if (invocationEvent is CompletionMessage) {
             if (invocationEvent.error != null) {
-              if (invocationEvent.error.isNotEmpty) {
+              if (invocationEvent.error!.isNotEmpty) {
                 controller.addError(Exception(error));
               } else {
                 controller.close();
@@ -539,14 +538,16 @@ class HubConnection {
           }
         } else {
           if (invocationEvent is StreamItemMessage) {
-            controller.add(invocationEvent.item as T);
+            controller.add(invocationEvent.item as T?);
           }
         }
       }
     };
 
     futureQueue = _sendWithProtocol(invocationDescriptor).catchError((e) {
-      controller.addError(e);
+      if (e is Object) {
+        controller.addError(e);
+      }
       _callbacks.remove(invocationDescriptor.invocationId);
     });
 
@@ -560,7 +561,7 @@ class HubConnection {
   /// The Future returned by this method resolves when the server indicates it has finished invoking the method. When the future
   /// resolves, the server has finished invoking the method. If the server method returns a result, it is produced as the result of
   /// resolving the Future.
-  Future<dynamic> invoke(String methodName, {List<dynamic> args}) {
+  Future<dynamic> invoke(String methodName, {List<dynamic>? args}) {
     final streamParameters = _replaceStreamParameters(args);
     final invocationDescriptor = _createInvocation(
       methodName: methodName,
@@ -572,7 +573,7 @@ class HubConnection {
     final p = Completer<dynamic>();
 
     _callbacks[invocationDescriptor.invocationId] =
-        (HubMessage invocationEvent, Exception error) {
+        (HubMessage? invocationEvent, Exception? error) {
       if (error != null) {
         p.completeError(error);
       } else if (invocationEvent != null) {
@@ -593,7 +594,9 @@ class HubConnection {
     };
 
     final futureQueue = _sendWithProtocol(invocationDescriptor).catchError((e) {
-      p.completeError(e);
+      if (e is Object) {
+        p.completeError(e);
+      }
       _callbacks.remove(invocationDescriptor.invocationId);
     });
 
@@ -602,9 +605,9 @@ class HubConnection {
     return p.future;
   }
 
-  Tuple2<Map<int, Stream<dynamic>>, List<String>> _replaceStreamParameters(
-      List<dynamic> args) {
-    final streams = <int, Stream<dynamic>>{};
+  Tuple2<Map<int?, Stream<dynamic>>, List<String>> _replaceStreamParameters(
+      List<dynamic>? args) {
+    final streams = <int?, Stream<dynamic>>{};
     final streamIds = <String>[];
 
     if (args != null) {
@@ -612,7 +615,9 @@ class HubConnection {
         final argument = args[i];
         if (argument is Stream) {
           final streamId = _invocationId;
-          _invocationId++;
+          if (_invocationId != null) {
+            _invocationId = _invocationId! + 1;
+          }
           // Store the stream for later use
           streams[streamId] = argument;
           streamIds.add(streamId.toString());
@@ -623,12 +628,12 @@ class HubConnection {
       }
     }
 
-    return Tuple2<Map<int, Stream<dynamic>>, List<String>>(streams, streamIds);
+    return Tuple2<Map<int?, Stream<dynamic>>, List<String>>(streams, streamIds);
   }
 
   /// Registers a handler that will be invoked when the hub method with the specified method name is invoked.
-  void on(String methodName, MethodInvacationFunc newMethod) {
-    if (methodName.isEmpty || newMethod == null) {
+  void on(String methodName, MethodInvocationFunc newMethod) {
+    if (methodName.isEmpty) {
       return;
     }
 
@@ -638,21 +643,21 @@ class HubConnection {
     }
 
     // Preventing adding the same handler multiple times.
-    if (_methods[methodName].contains(newMethod)) {
+    if (_methods[methodName]!.contains(newMethod)) {
       return;
     }
 
-    _methods[methodName].add(newMethod);
+    _methods[methodName]!.add(newMethod);
   }
 
   /// Removes all handlers for the specified hub method.
-  void off(String methodName, {MethodInvacationFunc method}) {
+  void off(String methodName, {MethodInvocationFunc? method}) {
     if (methodName.isEmpty) {
       return;
     }
 
     methodName = methodName.toLowerCase();
-    final handlers = _methods[methodName];
+    final List<void Function(List<dynamic>)>? handlers = _methods[methodName];
     if (handlers == null) {
       return;
     }
@@ -679,12 +684,12 @@ class HubConnection {
     }
 
     // Data may have all been read when processing handshake response
-    if (data != null) {
+    if (data != null && data is Object) {
       // Parse the messages
       final messages = _protocol.parseMessages(data, _logger);
 
       for (final message in messages) {
-        switch (message.type) {
+        switch (message!.type) {
           case MessageType.undefined:
             break;
           case MessageType.invocation:
@@ -692,12 +697,13 @@ class HubConnection {
             break;
           case MessageType.streamItem:
           case MessageType.completion:
-            var invovationMessage = message as HubInvocationMessage;
-            final callback = _callbacks[invovationMessage.invocationId];
+            var invocationMessage = message as HubInvocationMessage;
+            final void Function(HubMessage, Exception?)? callback =
+                _callbacks[invocationMessage.invocationId];
 
             if (callback != null) {
               if (message.type == MessageType.completion) {
-                _callbacks.remove(invovationMessage.invocationId);
+                _callbacks.remove(invocationMessage.invocationId);
               }
               callback(message, null);
             }
@@ -711,26 +717,27 @@ class HubConnection {
             break;
           case MessageType.close:
             final closeMessage = message as CloseMessage;
-            _logger(
+            _logger!(
                 LogLevel.information, 'Close message received from server.');
 
             final exception = (closeMessage.error != null)
                 ? Exception(
-                    'Server returned an error on close: ' + closeMessage.error)
+                    'Server returned an error on close: ' + closeMessage.error!)
                 : null;
 
             if (closeMessage.allowReconnect == true) {
               // It feels wrong not to await connection.stop() here, but processIncomingData is called as part of an onreceive callback which is not async,
               // this is already the behavior for serverTimeout(), and HttpConnection.Stop() should catch and log all possible exceptions.
 
-              _connection.stop(exception: exception);
+              _connection!.stop(exception: exception);
             } else {
               // We cannot await stopInternal() here, but subsequent calls to stop() will await this if stopInternal() is still ongoing.
               _stopFuture = _stopInternal(exception: exception);
             }
             break;
           default:
-            _logger(LogLevel.warning, 'Invalid message type: ${message.type}.');
+            _logger!(
+                LogLevel.warning, 'Invalid message type: ${message.type}.');
             break;
         }
       }
@@ -749,7 +756,7 @@ class HubConnection {
       responseMessage = response.item2;
     } catch (e) {
       final message = 'Error parsing handshake response: ' + e.toString();
-      _logger(LogLevel.error, message);
+      _logger!(LogLevel.error, message);
 
       final exception = Exception(message);
       _handshakeCompleter.completeError(exception);
@@ -758,22 +765,22 @@ class HubConnection {
 
     if (responseMessage.error != null) {
       final message =
-          'Server returned handshake error: ' + responseMessage.error;
-      _logger(LogLevel.error, message);
+          'Server returned handshake error: ' + responseMessage.error!;
+      _logger!(LogLevel.error, message);
 
       final exception = Exception(message);
       _handshakeCompleter.completeError(exception);
       throw exception;
     } else {
-      _logger(LogLevel.debug, 'Server handshake complete.');
+      _logger!(LogLevel.debug, 'Server handshake complete.');
     }
 
     _handshakeCompleter.complete();
     return remainingData;
   }
 
-  void _connectionClosed({Exception exception}) {
-    _logger(LogLevel.debug,
+  void _connectionClosed({Exception? exception}) {
+    _logger!(LogLevel.debug,
         'HubConnection.connectionClosed(${exception.toString()}) called while in state ${_connectionState.toString()}.');
 
     // Triggering this.handshakeRejecter is insufficient because it could already be resolved without the continuation having run yet.
@@ -815,7 +822,8 @@ class HubConnection {
   }
 
   void _cancelCallbacksWithError(Exception exception) {
-    final callbacks = _callbacks;
+    final Map<String?, void Function(HubMessage?, Exception)> callbacks =
+        _callbacks;
     _callbacks = {};
 
     callbacks.forEach((key, value) {
@@ -824,13 +832,13 @@ class HubConnection {
   }
 
   InvocationMessage _createInvocation({
-    String methodName,
-    List<dynamic> args,
-    bool nonblocking,
-    List<String> streamIds,
+    String? methodName,
+    List<dynamic>? args,
+    required bool nonblocking,
+    List<String>? streamIds,
   }) {
     if (nonblocking) {
-      if (streamIds.isNotEmpty) {
+      if (streamIds!.isNotEmpty) {
         return InvocationMessage(
           arguments: args,
           streamIds: streamIds,
@@ -844,9 +852,11 @@ class HubConnection {
       }
     } else {
       final invocationId = _invocationId;
-      _invocationId++;
+      if (_invocationId != null) {
+        _invocationId = _invocationId! + 1;
+      }
 
-      if (streamIds.isNotEmpty) {
+      if (streamIds!.isNotEmpty) {
         return InvocationMessage(
           arguments: args,
           invocationId: invocationId.toString(),
@@ -864,12 +874,14 @@ class HubConnection {
   }
 
   StreamInvocationMessage _createStreamInvocation({
-    String methodName,
-    List<dynamic> args,
-    List<String> streamIds,
+    String? methodName,
+    List<dynamic>? args,
+    required List<String> streamIds,
   }) {
     final invocationId = _invocationId;
-    _invocationId++;
+    if (_invocationId != null) {
+      _invocationId = _invocationId! + 1;
+    }
 
     if (streamIds.isNotEmpty) {
       return StreamInvocationMessage(
@@ -887,17 +899,17 @@ class HubConnection {
     }
   }
 
-  CancelInvocationMessage _createCancelInvocation({String id}) {
+  CancelInvocationMessage _createCancelInvocation({String? id}) {
     return CancelInvocationMessage(invocationId: id);
   }
 
-  StreamItemMessage _createStreamItemMessage({String id, dynamic item}) {
+  StreamItemMessage _createStreamItemMessage({String? id, dynamic item}) {
     return StreamItemMessage(invocationId: id, item: item);
   }
 
   CompletionMessage _createCompletionMessage({
-    String id,
-    Exception exception,
+    String? id,
+    Exception? exception,
     dynamic result,
   }) {
     if (exception != null) {
@@ -910,12 +922,12 @@ class HubConnection {
     return CompletionMessage(invocationId: id, result: result);
   }
 
-  void _launchStreams(Map<int, Stream> streams, Future<void> futureQueue) {
+  void _launchStreams(Map<int?, Stream> streams, Future<void> futureQueue) {
     if (streams.isEmpty) {
       return;
     }
 
-    futureQueue ??= Completer<void>().future;
+    futureQueue = Completer<void>().future;
 
     streams.forEach((streamId, stream) {
       stream.listen((data) {
@@ -950,23 +962,17 @@ class HubConnection {
 
   /// Registers a handler that will be invoked when the connection is closed.
   void onclose(ClosedCallback callback) {
-    if (callback != null) {
-      _closedCallbacks.add(callback);
-    }
+    _closedCallbacks.add(callback);
   }
 
   /// Registers a handler that will be invoked when the connection starts reconnecting.
   void onreconnecting(ReconnectingCallback callback) {
-    if (callback != null) {
-      _reconnectingCallbacks.add(callback);
-    }
+    _reconnectingCallbacks.add(callback);
   }
 
   /// Registers a handler that will be invoked when the connection successfully reconnects.
   void onreconnected(ReconnectedCallback callback) {
-    if (callback != null) {
-      _reconnectedCallbacks.add(callback);
-    }
+    _reconnectedCallbacks.add(callback);
   }
 
   void _cleanupPingTimer() {
@@ -981,13 +987,13 @@ class HubConnection {
     }
   }
 
-  void _clearTimeout(Timer timer) {
+  void _clearTimeout(Timer? timer) {
     timer?.cancel();
     timer = null;
   }
 
   void _invokeClientMethod(InvocationMessage invocationMessage) {
-    final methods = _methods[invocationMessage.target.toLowerCase()];
+    final methods = _methods[invocationMessage.target!.toLowerCase()];
 
     if (methods != null) {
       try {
@@ -995,21 +1001,21 @@ class HubConnection {
           method(invocationMessage.arguments);
         }
       } catch (e) {
-        _logger(LogLevel.error,
-            'A callback for the method ${invocationMessage.target.toLowerCase()} threw error \'${e.toString()}\'.');
+        _logger!(LogLevel.error,
+            'A callback for the method ${invocationMessage.target!.toLowerCase()} threw error \'${e.toString()}\'.');
       }
 
       if (invocationMessage.invocationId != null) {
         // This is not supported in v1. So we return an error to avoid blocking the server waiting for the response.
         final message =
             'Server requested a response, which is not supported in this version of the client.';
-        _logger(LogLevel.error, message);
+        _logger!(LogLevel.error, message);
 
         // We don't want to wait on the stop itself.
         _stopFuture = _stopInternal(exception: Exception(message));
       }
     } else {
-      _logger(LogLevel.warning,
+      _logger!(LogLevel.warning,
           'No client method with the name \'${invocationMessage.target}\' found.');
     }
   }
