@@ -22,7 +22,7 @@ enum MessageType {
   close, // = 7,
 }
 
-extension MessageTypeExtensions on MessageType {
+extension MessageTypeExtensions on MessageType? {
   int get value {
     switch (this) {
       case MessageType.undefined:
@@ -75,22 +75,22 @@ abstract class HubMessage {
   const HubMessage({this.type});
 
   /// A [MessageType] value indicating the type of this message.
-  final MessageType type;
+  final MessageType? type;
 }
 
 /// Defines properties common to all Hub messages relating to a specific invocation.
 abstract class HubInvocationMessage extends HubMessage {
-  HubInvocationMessage({MessageType type, this.headers, this.invocationId})
+  HubInvocationMessage({MessageType? type, this.headers, this.invocationId})
       : super(type: type);
 
   /// A [MessageHeaders] dictionary containing headers attached to the message.
-  final Map<String, String> headers;
+  final Map<String, String>? headers;
 
   ///The ID of the invocation relating to this message.
   ///
   ///This is expected to be present for StreamInvocationMessage and CompletionMessage. It may
   ///be 'undefined' for an InvocationMessage if the sender does not expect a response.
-  final String invocationId;
+  final String? invocationId;
 }
 
 /// A hub message representing a non-streaming invocation.
@@ -99,21 +99,21 @@ class InvocationMessage extends HubInvocationMessage {
       {this.target,
       this.arguments,
       this.streamIds,
-      Map<String, String> headers,
-      String invocationId})
+      Map<String, String>? headers,
+      String? invocationId})
       : super(
             type: MessageType.invocation,
             headers: headers,
             invocationId: invocationId);
 
   /// The target method name.
-  final String target;
+  final String? target;
 
   /// The target method arguments.
-  final List<dynamic> arguments;
+  final List<dynamic>? arguments;
 
   /// The target method stream IDs.
-  final List<String> streamIds;
+  final List<String>? streamIds;
 }
 
 /// A hub message representing a streaming invocation.
@@ -122,27 +122,27 @@ class StreamInvocationMessage extends HubInvocationMessage {
       {this.target,
       this.arguments,
       this.streamIds,
-      Map<String, String> headers,
-      String invocationId})
+      Map<String, String>? headers,
+      String? invocationId})
       : super(
             type: MessageType.streamInvocation,
             headers: headers,
             invocationId: invocationId);
 
   /// The target method name.
-  final String target;
+  final String? target;
 
   /// The target method arguments.
-  final List<dynamic> arguments;
+  final List<dynamic>? arguments;
 
   /// The target method stream IDs.
-  final List<String> streamIds;
+  final List<String>? streamIds;
 }
 
 /// A hub message representing a single item produced as part of a result stream.
 class StreamItemMessage extends HubInvocationMessage {
   StreamItemMessage(
-      {this.item, Map<String, String> headers, String invocationId})
+      {this.item, Map<String, String>? headers, String? invocationId})
       : super(
             type: MessageType.streamItem,
             headers: headers,
@@ -157,8 +157,8 @@ class CompletionMessage extends HubInvocationMessage with EquatableMixin {
   CompletionMessage(
       {this.error,
       this.result,
-      Map<String, String> headers,
-      String invocationId})
+      Map<String, String>? headers,
+      String? invocationId})
       : super(
             type: MessageType.completion,
             headers: headers,
@@ -167,7 +167,7 @@ class CompletionMessage extends HubInvocationMessage with EquatableMixin {
   /// The error produced by the invocation, if any.
   ///
   /// Either CompletionMessage.error CompletionMessage.result must be defined, but not both.
-  final String error;
+  final String? error;
 
   /// The result produced by the invocation, if any.
   ///
@@ -175,7 +175,7 @@ class CompletionMessage extends HubInvocationMessage with EquatableMixin {
   final dynamic result;
 
   @override
-  List<Object> get props => [error, result, headers, invocationId];
+  List<Object?> get props => [error, result, headers, invocationId];
 }
 
 /// A hub message indicating that the sender is still active.
@@ -183,7 +183,7 @@ class PingMessage extends HubMessage with EquatableMixin {
   PingMessage() : super(type: MessageType.ping);
 
   @override
-  List<Object> get props => [type];
+  List<Object?> get props => [type];
 }
 
 /// A hub message indicating that the sender is closing the connection.
@@ -197,16 +197,16 @@ class CloseMessage extends HubMessage {
   /// The error that triggered the close, if any.
   ///
   /// If this property is undefined, the connection was closed normally and without error.
-  final String error;
+  final String? error;
 
   /// If true, clients with automatic reconnects enabled should attempt to reconnect after receiving the CloseMessage.
   /// Otherwise, they should not.
-  final bool allowReconnect;
+  final bool? allowReconnect;
 }
 
 /// A hub message sent to request that a streaming invocation be canceled.
 class CancelInvocationMessage extends HubInvocationMessage {
-  CancelInvocationMessage({Map<String, String> headers, String invocationId})
+  CancelInvocationMessage({Map<String, String>? headers, String? invocationId})
       : super(
             type: MessageType.cancelInvocation,
             headers: headers,
@@ -218,13 +218,13 @@ abstract class HubProtocol {
   HubProtocol({this.name, this.version, this.transferFormat});
 
   /// The name of the protocol. This is used by SignalR to resolve the protocol between the client and server.
-  final String name;
+  final String? name;
 
   /// The version of the protocol.
-  final int version;
+  final int? version;
 
   /// The TransferFormat of the protocol. */
-  final TransferFormat transferFormat;
+  final TransferFormat? transferFormat;
 
   /// Creates an array of [HubMessage] objects from the specified serialized representation.
   ///
@@ -233,7 +233,7 @@ abstract class HubProtocol {
   /// [input] A string (json), or Uint8List (binary) containing the serialized representation.
   /// [Logger] logger A logger that will be used to log messages that occur during parsing.
 
-  List<HubMessage> parseMessages(Object input, Logging logging);
+  List<HubMessage?> parseMessages(Object input, Logging? logging);
 
   /// Writes the specified HubMessage to a string or ArrayBuffer and returns it.
   ///
