@@ -21,17 +21,18 @@ class JsonHubProtocol implements HubProtocol {
 
   /// Creates an array of [HubMessage] objects from the specified serialized representation.
   @override
-  List<HubMessage> parseMessages(dynamic input, Logging logging) {
+  List<HubMessage?> parseMessages(dynamic input, Logging? logging) {
     // Only JsonContent is allowed.
     if (!(input is String)) {
       throw Exception(
           'Invalid input for JSON hub protocol. Expected a string.');
     }
 
-    final jsonInput = input as String;
-    final hubMessages = <HubMessage>[];
+    final jsonInput = input;
+    final hubMessages = <HubMessage?>[];
 
-    if (input == null) {
+    // ignore: unnecessary_null_comparison
+    if (input != null) {
       return hubMessages;
     }
 
@@ -41,40 +42,40 @@ class JsonHubProtocol implements HubProtocol {
       final jsonData = json.decode(message);
       final messageType =
           _getMessageTypeFromJson(jsonData as Map<String, dynamic>);
-      HubMessage parsedMessage;
+      HubMessage? parsedMessage;
 
       switch (messageType) {
         case MessageType.invocation:
           parsedMessage = InvocationMessageExtensions.fromJson(
-              jsonData as Map<String, dynamic>);
+              jsonData);
           _isInvocationMessage(parsedMessage as InvocationMessage);
           break;
         case MessageType.streamItem:
           parsedMessage = StreamItemMessageExtensions.fromJson(
-              jsonData as Map<String, dynamic>);
+              jsonData);
           _isStreamItemMessage(parsedMessage as StreamItemMessage);
           break;
         case MessageType.completion:
           parsedMessage = CompletionMessageExtensions.fromJson(
-              jsonData as Map<String, dynamic>);
+              jsonData);
           _isCompletionMessage(parsedMessage as CompletionMessage);
           break;
         case MessageType.ping:
           parsedMessage =
-              PingMessageExtensions.fromJson(jsonData as Map<String, dynamic>);
+              PingMessageExtensions.fromJson(jsonData);
           // Single value, no need to validate
           break;
         case MessageType.close:
           parsedMessage =
-              CloseMessageExtensions.fromJson(jsonData as Map<String, dynamic>);
+              CloseMessageExtensions.fromJson(jsonData);
           // All optional values, no need to validate
           break;
         default:
           // Future protocol changes can add message types, old clients can ignore them
-          logging(
+          logging!(
               LogLevel.information,
               'Unknown message type \'' +
-                  parsedMessage.type.toString() +
+                  parsedMessage!.type.toString() +
                   '\' ignored.');
           continue;
       }
@@ -86,38 +87,31 @@ class JsonHubProtocol implements HubProtocol {
 
   /// Writes the specified [HubMessage] to a string and returns it.
   @override
-  String writeMessage(HubMessage message) {
+  String? writeMessage(HubMessage message) {
     switch (message.type) {
       case MessageType.undefined:
         break;
       case MessageType.invocation:
         return TextMessageFormat.write(
             json.encode((message as InvocationMessage).toJson()));
-        break;
       case MessageType.streamItem:
         return TextMessageFormat.write(
             json.encode((message as StreamItemMessage).toJson()));
-        break;
       case MessageType.completion:
         return TextMessageFormat.write(
             json.encode((message as CompletionMessage).toJson()));
-        break;
       case MessageType.streamInvocation:
         return TextMessageFormat.write(
             json.encode((message as StreamInvocationMessage).toJson()));
-        break;
       case MessageType.cancelInvocation:
         return TextMessageFormat.write(
             json.encode((message as CancelInvocationMessage).toJson()));
-        break;
       case MessageType.ping:
         return TextMessageFormat.write(
             json.encode((message as PingMessage).toJson()));
-        break;
       case MessageType.close:
         return TextMessageFormat.write(
             json.encode((message as CloseMessage).toJson()));
-        break;
       default:
         break;
     }
@@ -125,7 +119,7 @@ class JsonHubProtocol implements HubProtocol {
   }
 
   static MessageType _getMessageTypeFromJson(Map<String, dynamic> json) {
-    switch (json['type'] as int) {
+    switch (json['type'] as int?) {
       case 0:
         return MessageType.undefined;
       case 1:
@@ -186,12 +180,12 @@ class JsonHubProtocol implements HubProtocol {
 extension InvocationMessageExtensions on InvocationMessage {
   static InvocationMessage fromJson(Map<String, dynamic> json) {
     return InvocationMessage(
-      target: json['target'] as String,
+      target: json['target'] as String?,
       arguments:
-          (json['arguments'] as List)?.map((item) => item as Object)?.toList(),
-      headers: json['headers'] as Map<String, String>,
-      invocationId: json['invocationId'] as String,
-      streamIds: json['streamIds'] as List<String>,
+          (json['arguments'] as List?)?.map((item) => item as Object).toList(),
+      headers: json['headers'] as Map<String, String>?,
+      invocationId: json['invocationId'] as String?,
+      streamIds: json['streamIds'] as List<String>?,
     );
   }
 
@@ -222,8 +216,8 @@ extension StreamItemMessageExtensions on StreamItemMessage {
   static StreamItemMessage fromJson(Map<String, dynamic> json) {
     return StreamItemMessage(
       item: json['item'] as dynamic,
-      headers: json['headers'] as Map<String, String>,
-      invocationId: json['invocationId'] as String,
+      headers: json['headers'] as Map<String, String>?,
+      invocationId: json['invocationId'] as String?,
     );
   }
 
@@ -249,9 +243,9 @@ extension CompletionMessageExtensions on CompletionMessage {
   static CompletionMessage fromJson(Map<String, dynamic> json) {
     return CompletionMessage(
       result: json['result'],
-      error: json['error'] as String,
-      headers: json['headers'] as Map<String, String>,
-      invocationId: json['invocationId'] as String,
+      error: json['error'] as String?,
+      headers: json['headers'] as Map<String, String>?,
+      invocationId: json['invocationId'] as String?,
     );
   }
 
@@ -279,7 +273,7 @@ extension PingMessageExtensions on PingMessage {
 
 extension CloseMessageExtensions on CloseMessage {
   static CloseMessage fromJson(Map<String, dynamic> json) {
-    return CloseMessage(error: json['error'] as String);
+    return CloseMessage(error: json['error'] as String?);
   }
 
   Map<String, dynamic> toJson() {
