@@ -10,7 +10,7 @@ import 'package:web_socket_channel/web_socket_channel.dart';
 
 Future<WebSocketChannel> connect(
   Uri uri, {
-  BaseClient client,
+  required BaseClient client,
 }) async {
   var random = Random();
 
